Refactor setup sagas
Considering the pattern of delegating to a sub-saga, we want to spawn
instead of fork. If we fork a child inside a subsaga, the subsaga doesn't
complete until its children do.

If we spawn instead, then we can `call` a subsaga, rather than using
yield*, as recommended by the docs:
https://redux-saga.js.org/docs/advanced/ComposingSagas.html

In addition, we expect a `setupBlah` sub-saga to complete before moving
on, but as in this case, a part of that setup might be to kick off
a long running saga.