WIP: try to get opening tests to pass again
The issue at hand is that the wallet and the app need to be in sync with
which address will be used by the participant.

The current behaviour is:
1. The wallet, upon login, fetches a private key and address from
   firebase
2. The app asks the wallet for that address
3. The app puts that address in a preFundSetup commitment
4. The app tells the wallet about that commitment.

The behaviour going forward is:
1. The wallet will not store private keys in firebase. Instead, a new
   private key will be generated for each channel that's opened.

This does not define the synchronization method. There are two options:
a.
  i.   The app creates an address and private key.
  ii.  The app creates a commitment with that private key, and tells the wallet about that commitment.
  iii. In this message, the app includes the private key.
b.
  i.  The app creates an incomplete preFundSetup setup, with
  a placeholder address in its index.
  ii. The wallet generates a private key, replaces the placeholder with
  a new address, and tells the app about it. The app now knows its
  address (though it doesn't seem to need to ever use that information,
  and can disregard it.)

The app shouldn't be concerned with private keys, so b seems to be the
right choice.
This means we would have to create a new pair of actions:

```
export const OWN_PREFUND_SETUP_COMMITMENT_RECEIVED = 'WALLET.OWN_PREFUND_SETUP_COMMITMENT_RECEIVED';
export const ownPreFundSetupCommitmentReceived = (
  commitment: IncompleteCommitment,
) => ({
  ...
});
```
(and likewise one of type OPPONENT_PREFUND_SETUP_COMMITMENT_RECEIVED)