Extract shared branches of wallet state into interface
This partially reverses some of the work that Tom did in
58662d5ce014140860bfb0619df3d959858cd5a1

However, Tom's intention was to avoid having interface A extend B, which
extends C, which extends D, etc.

This is a middle ground, where interfaces A, B, C all extend Shared.

The Shared interface is intentionally not exported -- it is simply
a more maintainable way of creating wallet states that have some shared
properties.