Remove ProtocolPayload
The point of the protocol property was to determine how to route an
action that's meant to create a new process.
But it seems like every message that gets sent needs to include a
processId.
In this case, when a message might start a new process, the data should
determine what protocol to run.