Only test packages changed since master
This is slightly dangerous if we don't properly use semantic versioning,
since:
A. if rps depends on wallet version 1.0.*
B. and we update the wallet package to version 1.0.1, which includes
a breaking change

then rps tests would fail, but they won't be run on circle.

Note that lerna filters on packages changed since origin/master. For
some reason, circle's `checkout` command fast-forwards the `master`
branch to the branch being tested, and thus no packages are changed
since `master`.