Use yarn.lock in cache key for all packages
If
1. packages A and B depend on v0 of package C
2. yarn installs package C in apps/node_modules
3. package A upgrades to C@v1
4. yarn installs package c@v1 in apps/node_modules

then the current cache invalidation strategy does not work for package
B, and package b's yarn cache does not contain C@v0

In theory, `yarn install` should remedy this, but it doesn't seem to
work as expected with magmo-devtools. This may have to do with
magmo-devtools being installed from a github url.