Wallet tsc creates declaration files
This affects how the server package imports files from the wallet
package. In the absence of d.ts files, when compiling the server
package, which depends on the wallet package, the entire wallet package
is also compiled. This is not only slow, but doesn't work, since tsc is
incorrectly configured in the context of the server.