WalletProtocol only includes protocols running in a process
WalletProtocol sort of needs to be used by anyone running a wallet,
since it's used as part of the process id synchronization, and
communication.

Thus, it doesn't make sense for WalletProtocol to include a protocol
that's specific to eg. the react client wallet.

Since these protocols are never referenced, they're removed.