Generalized approach to ledger challenging
(Not longer using actions specific to the indirect-defunding protocol).
This should make any future work enabling challenges in indirect-funding
much easier.

- ledger-challenging.md outlines the new approach, which tries to follow
application challenging as far as possible
- there is now a new internalMessageOutbox with associated helpers and
actions. When in application challenge mode, we post to the usual
messageOutbox. When in a ledger challenge, we post to the new outbox.
- there is also a new saga internal-message-sender, which dispatches
actions instead of sending messages to the parent window
- challengingreducer renamed to disputeReducer
- isYieldingProcessApplication helper is used to switch reducers (so
they can decide, e.g. whether to hide the wallet or instead yield to
another wallet process)
- challenge-response-relay renamed to ledger-dispute-relay