Improve 'commitment sent' expecters
The old way assumed that the commitments sent were the first item in the
message outbox. This would fail if the correct message was the second
item in the message outbox.

The new way passes if at least one of the outbox items matches. If not,
it falls back to the old way in order to  produce readable diffs.