Use blocking call in rps saga
Because the preceding `take` is from an eventChannel, the saga is OK to block on the firebase API call, since any following events will be cached and then handled when the API call finishes.

Moreover, using `call` instead of `fork` will ensure the database is written to in the same order that the events are taken.